pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = ""
authors = ["Zach Mitchell <zmitchell@fastmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
molecule = {version = "^2.22",extras = ["docker"]}
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
